Concept 1 example
This is an example of division scape. A 2D math equations-complete game. In this game the user chooses an equal sign to complete an equation. It is primarily a clicking game which involves solving different types of division and multiplication problems. A close example apk can be found here: https://drive.google.com/open?id=11QAmBUJFL3neve2fCBbTlQwf-s8osNOd a prototype is in development and will be added soon